Move utilities to a utils file
